requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI & ML
weaviate-client==3.25.3
google-generativeai==0.3.2
langchain==0.0.352
langchain-community==0.0.8
sentence-transformers==2.2.2
numpy==1.24.4

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0

# Audio Processing (optional - install separately if needed)
# SpeechRecognition==3.10.0
# pydub==0.25.1
# gTTS==2.4.0
# pyaudio==0.2.11

# Background Tasks
celery==5.3.4
redis==5.0.1

# Email
email-validator==2.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
Jinja2==3.1.2

# Monitoring & Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
